refactor: remove validation of domain
