+++ util/validator.py
@@ -1,3 +1,4 @@
+import contextlib
 import re
 import socket
 
@@ -5,47 +6,27 @@ from prompt_toolkit.document import Document
 from questionary import ValidationError, Validator
 
 ipv4 = re.compile(r"((25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)(\.|$)){4}")
-domain = re.compile(
-    r"^[a-z]+([a-z0-9-]*[a-z0-9]+)?(\.([a-z]+([a-z0-9-]*[\[a-z0-9]+)?)+)*$"
-)
 
 
 class IPValidator(Validator):
-    def validate(self, document: Document):
+    def validate(self, document: Document) -> None:
         error = ValidationError(
-            message="Not a valid IP or URL", cursor_position=len(document.text)
+            message="Invalid IP", cursor_position=len(document.text)
         )
         try:
             ip = document.text
             if ipv4.match(ip):
                 socket.inet_aton(ip)
-            elif not domain.match(ip):
-                raise error
         except socket.error:
             raise error
 
     # TODO: Add an extra validator to check if an IP could be used by R* services (i.e. it's part of Microsoft Azure)
 
     @staticmethod
-    def validate_get(text: str):
-        error = ValidationError(
-            message="Not a valid IP or URL", cursor_position=len(text)
-        )
-        try:
-            ip = text
-            if ipv4.match(ip):
-                socket.inet_aton(ip)
-            elif domain.match(ip):
-                try:
-                    ip = socket.gethostbyname(text)
-                except socket.gaierror:
-                    raise ValidationError(
-                        message=f"URL {text} can't be resolved to IP",
-                        cursor_position=len(text),
-                    )
-                socket.inet_aton(ip)
-            else:
-                raise error
-            return ip
-        except socket.error:
-            raise error
+    def validate_get(text: str) -> str:
+        error = ValidationError(message="Invalid IP", cursor_position=len(text))
+        with contextlib.suppress(socket.error):
+            if ipv4.match(text):
+                socket.inet_aton(text)
+                return text
+        raise error
